Get Lucky version
In this version, the sketch is configured to play Get Lucky by Daft Punk. Then, it analyzes two frequency bands (bass and midrange) to generate a horizontal set of reactive arches that represent bass and a vertical set that represent midrange.
Might not be ideal to use with other songs, since it was configured for this song's kick and guitar frequencies.